Finished health potion effect ahhh
all three main features have been completed. Resynched rhythm so that it's timed properly with notes, added better health UI (had to get rid of lerp because it interfered with the health potion and couldn't figure it out). Added a completely new feature- going to add a fifth after the submission for fun!